[BREAKING] Move the CodeCommit and CodeBuild Actions from the codepipeline module into their @aws-cdk modules.
As agreed to with the CodePipeline team,
we want to move the concrete Action types from the codepipeline module
and into the modules of the resources they represent integrations with.
This is doing this refactoring for CodeCommit Repositories and CodeBuild Projects.
