+++ packages/@aws-cdk/codebuild/lib/index.ts
@@ -1,3 +1,4 @@
+export * from './pipeline-actions';
 export * from './project';
 export * from './source';
 export * from './artifacts';
