+++ packages/@aws-cdk/codebuild/package.json
@@ -43,6 +43,7 @@
   },
   "dependencies": {
     "@aws-cdk/codecommit": "^0.7.3-beta",
+    "@aws-cdk/codepipeline": "^0.7.3-beta",
     "@aws-cdk/core": "^0.7.3-beta",
     "@aws-cdk/events": "^0.7.3-beta",
     "@aws-cdk/iam": "^0.7.3-beta",
