+++ packages/@aws-cdk/codebuild/test/integ.pipeline-code-commit-build.expected.json
