+++ packages/@aws-cdk/codebuild/test/integ.pipeline-events.expected.json
