+++ packages/@aws-cdk/codecommit/lib/index.ts
@@ -1,4 +1,5 @@
-export * from './codecommit';
+export * from './repository';
+export * from './pipeline-action';
 
 // AWS::CodeCommit CloudFormation Resources:
 export * from './codecommit.generated';
