+++ packages/@aws-cdk/codepipeline/package.json
@@ -42,8 +42,6 @@
     "pkglint": "^0.7.3-beta"
   },
   "dependencies": {
-    "@aws-cdk/codebuild": "^0.7.3-beta",
-    "@aws-cdk/codecommit": "^0.7.3-beta",
     "@aws-cdk/core": "^0.7.3-beta",
     "@aws-cdk/events": "^0.7.3-beta",
     "@aws-cdk/iam": "^0.7.3-beta",
